+++ src/blueticks/_client.py
@@ -69,6 +69,7 @@ class Blueticks:
         from blueticks.resources.engines import EnginesResource
         from blueticks.resources.groups import GroupsResource
         from blueticks.resources.messages import MessagesResource
+        from blueticks.resources.newsletters import NewslettersResource
         from blueticks.resources.scheduled_messages import ScheduledMessagesResource
         from blueticks.resources.utils import UtilsResource
         from blueticks.resources.webhooks import WebhooksResource
@@ -81,6 +82,7 @@ class Blueticks:
         self.engines = EnginesResource(self)
         self.groups = GroupsResource(self)
         self.messages = MessagesResource(self)
+        self.newsletters = NewslettersResource(self)
         self.scheduled_messages = ScheduledMessagesResource(self)
         self.utils = UtilsResource(self)
         self.webhooks = WebhooksResource(self)