+++ src/blueticks/types/__init__.py
@@ -16,7 +16,8 @@ from blueticks.types.chats import (
     Participant,
 )
 from blueticks.types.groups import Group, GroupParticipant
-from blueticks.types.messages import Message
+from blueticks.types.messages import LinkPreview, Message, MediaKind, MessageType
+from blueticks.types.newsletters import Newsletter
 from blueticks.types.page import Page
 from blueticks.types.ping import Ping
 from blueticks.types.scheduled_messages import ScheduledMessage
@@ -37,10 +38,14 @@ __all__ = [
     "DeletedResource",
     "Group",
     "GroupParticipant",
+    "LinkPreview",
     "LoadOlderMessagesResponse",
+    "MediaKind",
     "MediaUrlResponse",
     "Message",
     "MessageAck",
+    "MessageType",
+    "Newsletter",
     "OkResponse",
     "Page",
     "Participant",